Allow users to set basicMode even if settingsUIReadOnly is true
Previously despite basicMode being editable on the settings screen it would not save when settingsUIReadOnly was true. The fact that it is grouped on the settings screen with the already editable and savable gitUserName and gitUserEmail settings suggests that basicMode is intended to be set by users and the inability to do is an oversight.