Fix lut files that do not exist as never being considered outdated for the purpose of importing
The timestamp returned by %RoutineMgr.TS() is always the current time for a lut file that does not exist or has yet to be imported, resulting in such lut files as never being considered as outdated when performing an import.

Check that lut files exist before using the returned timestamp, allowing them to be imported.