Fix Import All not importing anything that doesn't already exist when compileOnImport is not set
For an item that is not already in source control, Import All would add it to source control and then if compileOnImport was set the default pull handler would be invoked, which would result in the item being imported too. However, when compileOnImport is not set the default pull handler would not be invoked, and the item would not be imported.

This change makes it so that the item is imported when compileOnImport is not set.