Attempt to parse HL7 and LUT files to determine their internal name
`$system.OBJ.Load` is able to load CLS and INC files to determine their internal name. However it is not able to load the HL7 and LUT files exported by Embedded Git. These are XML files with the internal name readily available within, make an attempt to parse the file and get the internal name.

This will help avoid situations such as HL7 or LUT files being introduced at the same time as their Embedded Git mappings where the internal name was previously not able to be determined due to the mappings getting loaded in later and resulting in them not being loaded at all after a pull.